Refactor audio model training and configuration: remove obsolete files, update trainer imports, and enhance logging functionality.

+++ owl_audio/utils/logging.py
@@ -75,7 +75,7 @@ def log_audio_to_wandb(
     for i in range(batch_size):
         # Convert to numpy and ensure correct shape for wandb
         # (B, N, D) -> (N, D)
-        audio = sample[i].detach().cpu().numpy()  # (N, D)
+        audio = sample[i].detach().cpu().float().numpy()  # (N, D)
 
         # For stereo audio, mix down to mono for logging
         if audio.shape[-1] == 2:
